RR/requirements.txt
kafka-python==2.0.2
textblob==0.17.1
plyer==2.1.0
python-dotenv==1.0.1
nltk==3.8.1
